Move history to the separate endpoint
